Write a function that prints an integer

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,7 +7,7 @@
  */
 void print_number(int n)
 {
-	int x = n;
+	unsigned int x = n;
 
 	if (n < 0)
 	{
@@ -19,5 +19,5 @@ void print_number(int n)
 	if (x != 0)
 		print_number(x);
 
-	_putchar(n % 10 + '0');
+	_putchar((unsigned int) n % 10 + '0');
 }
